separate Mnemonic from Instruction
The idea here being that "Mnemonic" is the name and the associated behavior,
e.g. LDA and loading into the accumulator, while Instruction is a Mnemonic
bound into a program context with associated bytes to act upon. I think
this'll make it a lot easier once I properly parse the program RAM.