Issue #3: rewrite PPU scroll control to closely resemble hardware
This rewrites the previous model, where the PPU had two sets PPUScrollState
objects, to now be a single PPUInternalRegisters. Its fields are named like
the actual hardware's. A number of convenience methods are used so that the
caller can still think in terms of coarse/fine x/y when that's more
appropriate.

This does not go all the way, though: because the single render_scanline()
method does not closely align with the hardware, we need to have writes to
PPUADDR continue to be stored in a separate field, so that the CPU can
use them without worrying about the PPU clobbering them. This can go away
once render_scanline() is split up into smaller operations.

This also required the sprite zero detection to be rewritten, and it was
significantly simplified; it's now a part of the background tiles rendering.